migrate ignite.contrib.handlers
move loggers and other class from contrib to ignite.handlers
make changes in a backwards compatible manner
fix circular imports that happened during the migration in ignite/__init__.py
and ignite/handlers/__init__.py
