+++ docs/source/handlers.rst
@@ -1,8 +1,8 @@
 ignite.handlers
 ===============
 
-Complete list of handlers
--------------------------
+Complete list of generic handlers
+----------------------------------
 
 .. currentmodule:: ignite.handlers
 
@@ -32,6 +32,46 @@ Complete list of handlers
     checkpoint.BaseSaveHandler
     param_scheduler.ParamScheduler
     state_param_scheduler.StateParamScheduler
+
+
+Loggers
+--------
+
+.. currentmodule:: ignite.handlers
+
+.. autosummary::
+    :nosignatures:
+    :toctree: generated
+    :recursive:
+
+    base_logger
+    clearml_logger
+    mlflow_logger
+    neptune_logger
+    polyaxon_logger
+    tensorboard_logger
+    tqdm_logger
+
+    visdom_logger
+    wandb_logger
+
+.. seealso::
+
+    Below are a comprehensive list of examples of various loggers.
+
+    * See `tensorboardX mnist example <https://github.com/pytorch/ignite/blob/master/examples/mnist/mnist_with_tensorboard_logger.py>`_
+      and `CycleGAN and EfficientNet notebooks <https://github.com/pytorch/ignite/tree/master/examples/notebooks>`_ for detailed usage.
+
+    * See `visdom mnist example <https://github.com/pytorch/ignite/blob/master/examples/mnist/mnist_with_visdom_logger.py>`_ for detailed usage.
+
+    * See `neptune mnist example <https://github.com/pytorch/ignite/blob/master/examples/mnist/mnist_with_neptune_logger.py>`_ for detailed usage.
+
+    * See `tqdm mnist example <https://github.com/pytorch/ignite/blob/master/examples/mnist/mnist_with_tqdm_logger.py>`_ for detailed usage.
+
+    * See `wandb mnist example <https://github.com/pytorch/ignite/blob/master/examples/mnist/mnist_with_wandb_logger.py>`_ for detailed usage.
+
+    * See `clearml mnist example <https://github.com/pytorch/ignite/blob/master/examples/mnist/mnist_with_clearml_logger.py>`_ for detailed usage.
+
 
 .. _param-scheduler-label:
 
@@ -396,7 +436,7 @@ Example with :class:`ignite.handlers.param_scheduler.ReduceLROnPlateauScheduler`
     init_lr = 0.1
 
     lr_values = np.array(ReduceLROnPlateauScheduler.simulate_values(
-        num_events, metric_values, init_lr, 
+        num_events, metric_values, init_lr,
         factor=0.5, patience=1, mode='max', threshold=0.01, threshold_mode='abs'
         )
     )
