+++ ignite/__init__.py
@@ -1,3 +1,5 @@
+__version__ = "0.5.0"
+
 import ignite.contrib
 import ignite.distributed
 import ignite.engine
@@ -5,5 +7,3 @@ import ignite.exceptions
 import ignite.handlers
 import ignite.metrics
 import ignite.utils
-
-__version__ = "0.5.0"
